Rename hardware directories to not be lower-cased.
